Added cmake config 2

+++ modules/test_tasks/test_omp/ops.cpp
@@ -5,30 +5,30 @@
 #include <ctime>
 #include "../../../modules/test_tasks/test_omp/ops.h"
 
-std::vector<int> getRandomVector(size_t sz) {
+std::vector<int> getRandomVector(int  sz) {
     std::mt19937 gen;
     gen.seed(static_cast<unsigned int>(time(0)));
     std::vector<int> vec(sz);
-    for (size_t i = 0; i < sz; i++) { vec[i] = gen() % 100; }
+    for (int  i = 0; i < sz; i++) { vec[i] = gen() % 100; }
     return vec;
 }
 
 int getParallelOperations(std::vector<int> vec, std::string ops) {
-    const size_t sz = vec.size();
+    const int  sz = vec.size();
     int reduction_elem = 1;
     if (ops == "+") {
         #pragma omp parallel for reduction(+:reduction_elem)
-        for (size_t i = 0; i < sz; i++) {
+        for (int  i = 0; i < sz; i++) {
             reduction_elem += vec[i];
         }
     } else if (ops == "-") {
         #pragma omp parallel for reduction(-:reduction_elem)
-        for (size_t i = 0; i < sz; i++) {
+        for (int  i = 0; i < sz; i++) {
             reduction_elem -= vec[i];
         }
     } else if (ops == "*") {
         #pragma omp parallel for reduction(*:reduction_elem)
-        for (size_t i = 0; i < sz; i++) {
+        for (int  i = 0; i < sz; i++) {
             reduction_elem *= vec[i];
         }
     }
@@ -36,18 +36,18 @@ int getParallelOperations(std::vector<int> vec, std::string ops) {
 }
 
 int getSequentialOperations(std::vector<int> vec, std::string ops) {
-    const size_t sz = vec.size();
+    const int  sz = vec.size();
     int reduction_elem = 1;
     if (ops == "+") {
-        for (size_t i = 0; i < sz; i++) {
+        for (int  i = 0; i < sz; i++) {
             reduction_elem += vec[i];
         }
     } else if (ops == "-") {
-        for (size_t i = 0; i < sz; i++) {
+        for (int  i = 0; i < sz; i++) {
             reduction_elem -= vec[i];
         }
     } else if (ops == "*") {
-        for (size_t i = 0; i < sz; i++) {
+        for (int  i = 0; i < sz; i++) {
             reduction_elem *= vec[i];
         }
     }
